// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/claim-management/shared/DTOs/claim-document-received-report.dto.ts
export class ClaimDocumentReceivedReport_DTO {
    HospitalNo: number = null;
    PatientName: string = '';
    MemberNo: number = null;
    ClaimCode: number = null;
    InvoiceNo: string = '';
    InvoiceDate: string = '';
    TotalAmount: number = null;
    Scheme: string = '';
    DocumentReceivedDate: string = '';
    Remarks: string = '';
    User: string = '';
}